Move root actions file to SubmissionSection directory
All actions within this file are now SubmissionSection-specific, so no need for an Entity Editor-wide actions file.
